backend/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.0
charset-normalizer==3.0.1
click==8.1.3
coloredlogs==15.0.1
Flask==2.2.3
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
mpmath==1.2.1
numpy==1.24.2
oauthlib==3.2.2
onnx==1.12.0
onnxruntime==1.13.1
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sympy==1.11.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.14.1
flask-cors==3.0.10